requirements.txt
fastapi
uvicorn[standard]
openai
google-generativeai
google-ai-generativelanguage
google-cloud-aiplatform